+++ src/components/task_manager/TMTask.tsx
@@ -1,20 +1,8 @@
-import { Fragment, useState } from "react";
+import { useState } from "react";
 import { useTasks } from "../../services/provider/TasksProvider";
 import Task from "../../types/Task";
-import {
-  Button,
-  FormControl,
-  InputLabel,
-  MenuItem,
-  OutlinedInput,
-  Select,
-  Stack,
-  TextField,
-  Typography,
-} from "@mui/material";
-import { DatePicker, LocalizationProvider } from "@mui/x-date-pickers/";
-import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
-import dayjs from "dayjs";
+import { Button, Stack, Typography } from "@mui/material";
+
 import { useUsers } from "../../services/provider/UsersProvider";
 import Status from "../../types/Status";
 import { ItemTMTask } from "../../style/style";
@@ -23,6 +11,7 @@ import {
   TaskActionType,
 } from "../../types/providers-types/TasksProviderTypes";
 import { IUserContext } from "../../types/providers-types/UsersProviderTypes";
+import TMEditedTask from "./TMEditedTask";
 
 interface TMTaskProps {
   task: Task;
@@ -37,97 +26,20 @@ function TMTask({ task }: TMTaskProps) {
     return <></>;
   }
 
-  let usersOptions: JSX.Element[] = [];
-  for (let i = 0; i < users.users.length; i++) {
-    usersOptions.push(
-      <MenuItem key={users.users[i].id} value={users.users[i].name}>
-        {users.users[i].name}
-      </MenuItem>
-    );
-  }
-  let editedTask: Task = new Task(
-    task.id,
-    task.title,
-    task.description,
-    task.status,
-    task.assignedTo,
-    task.dueDate
-  );
-
-  let status: string | undefined = Object.keys(Status).find(
+  let statusKey: string | undefined = Object.keys(Status).find(
     (key: string) => Status[key].name === task.status
   );
 
   if (isEditing) {
     return (
-      <ItemTMTask>
-        <Typography variant="h6" color="primary">
-          Aggiorna la task
-        </Typography>
-        <TextField
-          id="outlined-basic"
-          label="Inserisci il titolo"
-          size="medium"
-          fullWidth
-          margin="normal"
-          value={task.title}
-          onChange={(e) => {
-            editedTask.title = e.target.value;
-            tasks.dispatch({
-              type: TaskActionType.EDIT,
-              task: editedTask,
-            });
-          }}
-        />
-        <TextField
-          id="outlined-basic"
-          label="Inserisci la descrizione"
-          size="medium"
-          fullWidth
-          margin="normal"
-          value={task.description}
-          onChange={(e) => {
-            editedTask.description = e.target.value;
-            tasks.dispatch({
-              type: TaskActionType.EDIT,
-              task: editedTask,
-            });
-          }}
-        />
-        {Status[status!].icon}
-        <FormControl fullWidth margin="normal">
-          <InputLabel id="demo-multiple-name-label">Utenti</InputLabel>
-          <Select
-            labelId="demo-multiple-name-label"
-            id="demo-multiple-name"
-            input={<OutlinedInput label="Name" />}
-            value={editedTask.assignedTo}
-            onChange={(e) => {
-              editedTask.assignedTo = e.target.value;
-              tasks.dispatch({
-                type: TaskActionType.EDIT,
-                task: editedTask,
-              });
-            }}
-          >
-            {usersOptions}
-          </Select>
-        </FormControl>
-        <LocalizationProvider dateAdapter={AdapterDayjs}>
-          <DatePicker
-            label="Scadenza"
-            value={dayjs(task.dueDate)}
-            onChange={(newDate) => {
-              editedTask.dueDate = newDate!.toISOString().slice(0, 10);
-              tasks.dispatch({
-                type: TaskActionType.EDIT,
-                task: editedTask,
-              });
-            }}
-          />
-        </LocalizationProvider>
-        <Button onClick={() => setIsEditing(!isEditing)}>Save</Button>
-      </ItemTMTask>
+      <TMEditedTask
+        task={task}
+        dispatch={tasks.dispatch}
+        statusKey={statusKey}
+        users={users.users}
+        isEditing={isEditing}
+        setIsEditing={setIsEditing}
+      ></TMEditedTask>
     );
   } else {
     return (
@@ -160,7 +72,7 @@ function TMTask({ task }: TMTaskProps) {
           {task.dueDate}
         </h3>
         <br />
-        <div style={{ marginTop: "-1.5rem" }}>{Status[status!].icon}</div>
+        <div style={{ marginTop: "-1.5rem" }}>{Status[statusKey!].icon}</div>
         <Button onClick={() => setIsEditing(!isEditing)}>Edit</Button>
         <Button
           onClick={() =>
